# api/openapi.yaml
openapi: 3.0.3
info:
  title: Robot selection API
  version: 0.6.0
paths:
...
  /api/solutions:
    get:
      summary: List catalog solutions
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: kind
          schema:
            type: string
        - in: query
          name: subtype
          schema:
            type: string
        - in: query
          name: vendor
          schema:
            type: string
        - in: query
          name: object_type
          schema:
            type: string
            enum: [warehouse, airport, hospital]
        - in: query
          name: min_payload_kg
          schema:
            type: number
        - in: query
          name: max_width_mm
          schema:
            type: number
        - in: query
          name: family
          description: Comma list of robot groups; a robot with no group counts as other.
          schema:
            type: string
            example: mobile,uav
        - in: query
          name: status
          description: Comma list of status codes (operation, piloting, rnd).
          schema:
            type: string
        - in: query
          name: specs_complete
          description: Only robots whose specs cover what matching needs.
          schema:
            type: boolean
        - in: query
          name: archived
          description: include adds archived robots (admin only, 403 otherwise).
          schema:
            type: string
            enum: [include]
        - in: query
          name: sort
          schema:
            type: string
            enum: [name, price_rub, payload_kg]
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            maximum: 500
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Solution"
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
  /api/solutions/{solutionId}:
    get:
      summary: One catalog solution
      parameters:
        - in: path
          name: solutionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solution"
        "404":
          description: Not found
  /api/solutions/{solutionId}/image:
    get:
      summary: A robot's photo (public)
      description: >-
        JPEG, or PNG for a photo with transparency, at most 1 200 px on the longer side. The ETag is the photo's
        sha256; a request with `v` equal to it is cached for a year.
      parameters:
        - $ref: "#/components/parameters/SolutionID"
        - in: query
          name: v
          schema:
            type: string
      responses:
        "200":
          description: The photo
          content:
            image/jpeg: {}
            image/png: {}
        "304":
          description: Not modified
        "404":
          description: The robot has no photo
  /api/catalog/bundle:
    get:
      summary: The live catalog as the calculation engine takes it
      description: The ETag is the catalog's content hash; a request with a matching If-None-Match gets 304.
      responses:
        "200":
          description: Catalog bundle
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogBundle"
        "304":
          description: Not modified
...
  /api/admin/norms:
    get:
      summary: Current calculation norms (admin)
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Norms
        "403":
          description: Not an admin
    put:
      summary: Replace calculation norms (admin)
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Saved
        "400":
          description: Out of range
        "403":
          description: Not an admin
...
  /api/object-types/{type}/schema:
    get:
      summary: Parameter schema for an object type
      parameters:
        - in: path
          name: type
          required: true
          schema:
            type: string
            enum: [warehouse, airport, hospital]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectSchema"
        "404":
          description: Unknown object type
  /api/guest/calculate:
    post:
      summary: Guest calculate (not persisted)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [object_type, params]
              properties:
                object_type:
                  type: string
                  enum: [warehouse, airport, hospital]
                params:
                  type: object
                  additionalProperties: true
                seed:
                  type: integer
                  minimum: 0
                include_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                overrides:
                  $ref: "#/components/schemas/EconOverrides"
                schema_version:
                  type: integer
                  description: Operation schema version of collections. Required with collections, otherwise 409 schema_version.
                collections:
                  type: object
                  additionalProperties: true
                  description: |
                    The project as operation collections, as in GET /api/projects/{id}/snapshot. When present it replaces
                    params: the processes, variants with their fleets, financing, assumptions, shared costs and the
                    what-if come from it. It is checked as an import is and used once; nothing is stored.
      responses:
        "200":
          description: Structured result with econ, match, model econ-v3. Not persisted.
        "400":
          description: Validation error, or collections that do not pass the operation schema
        "409":
          description: Unsupported schema version (code schema_version)
...
  /api/guest/match:
    post:
      summary: Guest matching (not persisted)
      parameters:
        - $ref: "#/components/parameters/MatchView"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [object_type, params]
              properties:
                object_type:
                  type: string
                  enum: [warehouse, airport, hospital]
                params:
                  type: object
                  additionalProperties: true
                include_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                task_codes:
                  type: array
                  items:
                    type: string
                    enum: [pallet_inbound, pallet_putaway, pallet_outbound, pallet_move, piece_pick, cleaning]
      responses:
        "200":
          description: Ranked catalog with hard, soft, missing_evidence and explanation chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchOutput"
        "400":
          description: Validation error
  /api/auth/register:
    post:
      summary: Register a user and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
                  minLength: 8
                  maxLength: 72
                invite_token:
                  type: string
                  description: Token from an invitation letter. Required when REGISTRATION_MODE is invite.
      responses:
        "200":
          description: Session started; the cookie is in Set-Cookie
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Invalid input
        "403":
          description: Cross-site request (code csrf), or registration is by invitation only (code invite_required)
        "409":
          description: Email already registered
  /api/auth/login:
    post:
      summary: Log in and start a session
      description: A session cookie already on the request is revoked and replaced.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Session started; the cookie is in Set-Cookie
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Bad credentials (the same text for an unknown email)
        "403":
          description: Cross-site request (code csrf)
  /api/auth/me:
    get:
      summary: Current user and the session's CSRF token
      security:
        - sessionCookie: []
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Not signed in, or the session expired (code session_expired)
  /api/auth/logout:
    post:
      summary: Revoke the current session and clear the cookie
      description: Succeeds without a session too.
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Logged out
        "403":
          description: Missing CSRF token (code csrf)
  /api/auth/verify-email:
    post:
      summary: Confirm an email address with the token from its letter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Confirmed
        "400":
          description: No token
        "410":
          description: The token is used or expired
  /api/auth/verify-email/resend:
    post:
      summary: Send the confirmation letter again
      security:
        - sessionCookie: []
      responses:
        "200":
          description: The address is already confirmed
        "202":
          description: Letter queued
        "401":
          description: Not signed in
  /api/auth/password-reset/request:
    post:
      summary: Ask for a password link
      description: Always answers the same way, so it never says which addresses exist.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
      responses:
        "202":
          description: Accepted
  /api/auth/password-reset/confirm:
    post:
      summary: Set a new password with the token from the letter
      description: Every session of that account ends, including the caller's.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 8
                  maxLength: 72
      responses:
        "200":
          description: Password changed
        "400":
          description: Invalid input
        "410":
          description: The token is used or expired
  /api/auth/invitations/inspect:
    post:
      summary: What an invitation token stands for, for the registration page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: The invitation, or why it cannot be used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationCheck"
        "400":
          description: No token
  /api/dev/outbox:
    get:
      summary: Letters queued for one address (only when APP_ENV is local)
      parameters:
        - name: to
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Letters, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/OutboxLetter"
        "404":
          description: Not served outside a local run
  /api/auth/sessions:
    get:
      summary: Active sessions of the current user, most recently used first
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        "401":
          description: Not signed in
  /api/auth/sessions/{sessionId}:
    delete:
      summary: Revoke one of the current user's sessions
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Revoked; the cookie is cleared when it was the current session
        "401":
          description: Not signed in
        "403":
          description: Missing CSRF token (code csrf)
        "404":
          description: No such active session of this user
  /api/auth/sessions/revoke-others:
    post:
      summary: Revoke every session of the current user except this one
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Number of revoked sessions
          content:
            application/json:
              schema:
                type: object
                required: [revoked]
                properties:
                  revoked:
                    type: integer
        "401":
          description: Not signed in
        "403":
          description: Missing CSRF token (code csrf)
  /api/guest/export:
    post:
      summary: Guest PDF or XLSX (not persisted)
      description: >-
        Financial report for a decision maker, built from the request and stored nowhere. Same content as
        POST /api/calculations/{runId}/export, without a run: no X-Run-Id header, the project is named "Не сохранён",
        the file is named ocenka-warehouse.xlsx after the object type.
      parameters:
        - in: query
          name: format
          schema:
            type: string
            enum: [pdf, xlsx]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [object_type, params]
              properties:
                object_type:
                  type: string
                  enum: [warehouse, airport, hospital]
                params:
                  type: object
                  additionalProperties: true
                seed:
                  type: integer
                  minimum: 0
                include_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                overrides:
                  $ref: "#/components/schemas/EconOverrides"
                schema_version:
                  type: integer
                  description: Operation schema version of collections. Required with collections, otherwise 409 schema_version.
                collections:
                  type: object
                  additionalProperties: true
                  description: |
                    The project as operation collections, as in GET /api/projects/{id}/snapshot. When present it replaces
                    params: the processes, variants with their fleets, financing, assumptions, shared costs and the
                    what-if come from it. It is checked as an import is and used once; nothing is stored.
      responses:
        "200":
          description: >-
            PDF (up to two pages for a typical project) or XLSX with the sheets Итог, Варианты, Допущения и риски.
            Money is numeric in XLSX and shown with the ruble sign.
        "400":
          description: Validation error, or collections that do not pass the operation schema
        "409":
          description: Unsupported schema version (code schema_version)
  /api/projects:
    get:
      summary: List own projects and projects shared with the caller
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Items carry access (owner, editor or viewer)
        "403":
          description: Guest cannot list projects
    post:
      summary: Create a project (owner)
      security:
        - sessionCookie: []
      responses:
        "201":
          description: Created
        "400":
          description: Invalid input
        "403":
          description: Guest cannot persist a project
  /api/projects/trash:
    get:
      summary: Own projects in the trash, newest first
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Trash
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/TrashItem"
        "401":
          description: Not signed in
...
  /api/projects/{id}/restore:
    post:
      summary: Take own project out of the trash (owner)
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The restored project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found, or not in the trash
        "410":
          description: Deleted longer ago than the retention
...
components:
  parameters:
    SolutionID:
      in: path
      name: solutionId
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-session
      description: >
        HttpOnly session cookie set by login and register. With COOKIE_SECURE=false
        (local only) the cookie is named `session`. POST, PUT, PATCH and DELETE under
        /api/ with a session also need the `X-CSRF-Token` header equal to `csrf_token`
        from login, register or /api/auth/me, and must come from the same origin
        (`Sec-Fetch-Site: same-origin`, or `Origin` equal to PUBLIC_ORIGIN); otherwise
        403 with code `csrf`. An expired or revoked cookie gives 401 with code
        `session_expired` and clears the cookie.
  schemas:
...
    EconOverrides:
      type: object
      properties:
        price_rub:
          type: number
        volume_factor:
          type: number
        labor_factor:
          type: number
        fleet_size:
          type: integer
        capex_rub:
          type: number
        opex_year_rub:
          type: number
...
    CatalogBundle:
      type: object
      required: [content_sha256, candidates, robots]
      properties:
        content_sha256:
          type: string
        candidates:
          type: array
          items:
            type: object
            additionalProperties: true
        robots:
          type: array
          items:
            type: object
            additionalProperties: true
...